Change weight type to float

+++ src/ExpressTransitionStrategy.java
@@ -1,6 +1,6 @@
 public class ExpressTransitionStrategy implements TransitionStrategy{
     @Override
-    public double getPrice(int weight) {
+    public double getPrice(float weight) {
         return 3.5 * weight;
     }
 }
